Reimplement header to fix header shrinking
Previously, the header and the hero section were both in the same flex
container and flex-direction: column was used to stack them on top of
each other. This, however, stopped the empty space in the header (the
space between the header logo and the header links) from shrinking when
the window became smaller. In order to fix this, the header was put into
a separate container, and the hero section was put into a separate
container, too.