Make hero section centered and resize dynamically
Before, there were widths set on .hero-left and .hero-image. This made
the two items not shrink (need to investigate behavior). By changing it
from width to flex-basis, shrinking was enabled.

Also, the items were center aligned vertically. This is meant to center
the image when .hero-left grows in height to wrap text when resizing.